requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.1.0
av==14.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
configparser==7.1.0
contourpy==1.3.1
ctranslate2==4.4.0
cycler==0.12.1
demucs==4.0.1
docopt==0.6.2
dora_search==0.1.12
einops==0.8.1
faster-whisper==1.1.0
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
huggingface-hub==0.28.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
lameenc==1.8.1
lightning==2.5.0.post0
lightning-utilities==0.12.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.25.2
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.20.1
openunmix==1.3.0
optuna==4.2.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primePy==1.3
propcache==0.2.1
protobuf==5.29.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semver==3.0.4
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.2
SQLAlchemy==2.0.38
streamlit==1.42.0
streamlit_change_language==0.0.2
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
treetable==0.2.5
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
whisperx==3.3.1
yarl==1.18.3
